packages/shared/pyproject.toml
[project]
name = "nexus-shared"
version = "0.1.0"
description = "Nexus shared types and utilities"
requires-python = ">=3.12,<3.13"
license = { file = "../../LICENSE-BSL" }
dependencies = [
    "pydantic>=2.9.0",
    "python-ulid>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]
